refactor stopwatch, move formatTime to helper file
